Build: (GE) Add gradle android cache fix plugin
Some Android plugin versions have issues with Gradle's build cache
feature. When applied to an Android project this plugin applies
workarounds for these issues based on the Android plugin and Gradle
versions.

Because the project is not using convention plugins, a quick way of
testing this plugin was by placing it in the root project's
'build.gradle' file.

For more details see:
https://github.com/gradle/android-cache-fix-gradle-plugin