Build: (GE) Build: Update agp to 7.1.3
This is done because by just applying the 'org.gradle.android.cache-fix'
plugin didn't seem to solve the 4 x 'CompileLibraryResourcesTask'
cacheable issues that we were after. Thus, and since this plugin
supports and is tested against the latest patch versions of each minor
version of AGP, maybe updating from '7.1.1' to the explicitly
supported by the plugin '7.1.3' version would do the trick.